+++ src/cavalry/test-ui.js
@@ -1,3 +1,4 @@
 const label = new ui.Label("Sent from VSCode")
 ui.add(label)
 ui.show()
+console.log('scriptLocation', ui.scriptLocation)