+++ src/commands/post.ts
@@ -1,4 +1,5 @@
 import { workspace, window } from 'vscode'
+import { temporaryWrite } from 'tempy'
 import axios from 'axios'
 
 export default async () => {
@@ -6,11 +7,9 @@ export default async () => {
 		.getConfiguration('stallion')
 		.get('address', '127.0.0.1')
 	const port = workspace.getConfiguration('stallion').get('port', 8080)
-
 	const server = `http://${address}:${port}/post`
-	console.log(server)
-
 	const doc = window.activeTextEditor?.document
+	console.log(server)
 
 	if (!doc) {
 		return window.showErrorMessage('No active document')
@@ -21,11 +20,18 @@ export default async () => {
 	}
 
 	try {
-		const filePath = doc.uri.fsPath
-		console.log(filePath)
-		const script = doc.getText()
-		const result = await axios.post(server, { data: script })
-		console.log(result)
+		const text = doc.getText()
+		const script = `(function() { ${text} })()`
+		const saved = !doc.isUntitled
+		const showUI = text.includes('ui.show()')
+		let path = await temporaryWrite(script, { extension: 'js' })
+		if (saved && showUI) {
+			path = doc.fileName
+		}
+		if (text.includes('ui.add(') && !showUI) {
+			window.showWarningMessage('Script is missing `ui.show()`')
+		}
+		const result = await axios.post(server, { path })
 		if (result?.statusText === 'OK') {
 			window.showInformationMessage('Successfully sent to Cavalry')
 		} else {